+++ src/App.tsx
@@ -1,5 +1,5 @@
-import DataPage from "./seamark/DataPage";
+import Deck from "./seamark/Deck";
 
 export default function App() {
-  return <DataPage />;
+  return <Deck />;
 }